Fix S2S message exchange in write_stanza(). XMLNS issues.
Tested with Psi and Google Chat.

write_stanza attempted to strip the jabber:server or jabber:client NS
from the XML so that the write would use the default connection XMLNS.
The strip was not good enough to understand the new XMLNS changes.

The fix is to convert the XML to text, and use RegExp to substitute the
correct jabber:server or jabber:client NS into the XML.

This is crude - It would be better to implement a proper XML NS
substitution routine.